# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgroGestion API",
    "description": "API para gestionar cultivos de principio a fin: siembras, ciclos, conteo de plantas, vivero, riesgos y eventos adversos. Todos los errores usan el mismo formato: decida siempre con el campo `error`, nunca con `message`.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/propagacion": {
      "get": {
        "tags": [
          "6. Vivero"
        ],
        "summary": "Lotes de vivero",
        "operationId": "listar_api_v1_propagacion_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "pendientes",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pendientes"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Registros a saltar",
              "default": 0,
              "title": "Skip"
            },
            "description": "Registros a saltar"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "description": "Tamaño de página, máximo 50",
              "default": 10,
              "title": "Limit"
            },
            "description": "Tamaño de página, máximo 50"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_PropagacionSalida_"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o sesión vencida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "403": {
            "description": "El rol no tiene permiso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "404": {
            "description": "No existe o no pertenece al usuario.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "409": {
            "description": "Choca con algo que ya existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "422": {
            "description": "Datos inválidos o regla de negocio.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "6. Vivero"
        ],
        "summary": "Crear un lote de vivero",
        "operationId": "crear_api_v1_propagacion_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PropagacionCrear"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropagacionSalida"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o sesión vencida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "403": {
            "description": "El rol no tiene permiso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "404": {
            "description": "No existe o no pertenece al usuario.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "409": {
            "description": "Choca con algo que ya existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "422": {
            "description": "Datos inválidos o regla de negocio.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/propagacion/{lote_id}": {
      "patch": {
        "tags": [
          "6. Vivero"
        ],
        "summary": "Actualizar germinadas, listas y pérdidas",
        "operationId": "actualizar_api_v1_propagacion__lote_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lote_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Lote Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PropagacionActualizar"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropagacionSalida"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o sesión vencida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "403": {
            "description": "El rol no tiene permiso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "404": {
            "description": "No existe o no pertenece al usuario.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "409": {
            "description": "Choca con algo que ya existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "422": {
            "description": "Datos inválidos o regla de negocio.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/propagacion/{lote_id}/trasplante": {
      "post": {
        "tags": [
          "6. Vivero"
        ],
        "summary": "Pasar plantas listas a una siembra",
        "operationId": "trasplantar_api_v1_propagacion__lote_id__trasplante_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lote_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Lote Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrasplanteEntrada"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropagacionSalida"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o sesión vencida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "403": {
            "description": "El rol no tiene permiso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "404": {
            "description": "No existe o no pertenece al usuario.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "409": {
            "description": "Choca con algo que ya existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "422": {
            "description": "Datos inválidos o regla de negocio.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/noticias": {
      "get": {
        "tags": [
          "11. Noticias"
        ],
        "summary": "Noticias vigentes de mi región",
        "operationId": "listar_api_v1_noticias_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "region",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(\\d{2}|\\d{5})$"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Código DANE; por defecto, el de sus fincas",
              "title": "Region"
            },
            "description": "Código DANE; por defecto, el de sus fincas"
          },
          {
            "name": "cultivo_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cultivo Id"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Registros a saltar",
              "default": 0,
              "title": "Skip"
            },
            "description": "Registros a saltar"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "description": "Tamaño de página, máximo 50",
              "default": 10,
              "title": "Limit"
            },
            "description": "Tamaño de página, máximo 50"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_NoticiaSalida_"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o sesión vencida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "403": {
            "description": "El rol no tiene permiso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "404": {
            "description": "No existe o no pertenece al usuario.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "409": {
            "description": "Choca con algo que ya existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "422": {
            "description": "Datos inválidos o regla de negocio.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "11. Noticias"
        ],
        "summary": "Cargar una noticia",
        "operationId": "crear_api_v1_noticias_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoticiaCrear"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoticiaSalida"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o sesión vencida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "403": {
            "description": "El rol no tiene permiso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "404": {
            "description": "No existe o no pertenece al usuario.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "409": {
            "description": "Choca con algo que ya existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "422": {
            "description": "Datos inválidos o regla de negocio.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DetalleError": {
        "properties": {
          "campo": {
            "type": "string",
            "title": "Campo"
          },
          "mensaje": {
            "type": "string",
            "title": "Mensaje"
          },
          "tipo": {
            "type": "string",
            "title": "Tipo"
          }
        },
        "type": "object",
        "required": [
          "campo",
          "mensaje",
          "tipo"
        ],
        "title": "DetalleError"
      },
...
      "ErrorRespuesta": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "status_code": {
            "type": "integer",
            "title": "Status Code"
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/DetalleError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "error",
          "message",
          "status_code"
        ],
        "title": "ErrorRespuesta",
        "description": "Formato único de error: decida siempre con `error`, nunca con `message`."
      },
...
      "NoticiaCrear": {
        "properties": {
          "titulo": {
            "type": "string",
            "maxLength": 200,
            "minLength": 3,
            "title": "Titulo"
          },
          "resumen": {
            "type": "string",
            "maxLength": 400,
            "minLength": 3,
            "title": "Resumen"
          },
          "enlace": {
            "type": "string",
            "maxLength": 400,
            "minLength": 8,
            "pattern": "^https?://",
            "title": "Enlace"
          },
          "fuente": {
            "type": "string",
            "maxLength": 120,
            "minLength": 2,
            "title": "Fuente"
          },
          "region_dane": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region Dane"
          },
          "cultivo_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cultivo Id"
          },
          "publicada": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publicada"
          },
          "vigente_hasta": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vigente Hasta"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "titulo",
          "resumen",
          "enlace",
          "fuente"
        ],
        "title": "NoticiaCrear"
      },
      "NoticiaSalida": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "titulo": {
            "type": "string",
            "title": "Titulo"
          },
          "resumen": {
            "type": "string",
            "title": "Resumen"
          },
          "enlace": {
            "type": "string",
            "title": "Enlace"
          },
          "fuente": {
            "type": "string",
            "title": "Fuente"
          },
          "region_dane": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region Dane"
          },
          "cultivo_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cultivo Id"
          },
          "publicada": {
            "type": "string",
            "format": "date-time",
            "title": "Publicada"
          },
          "vigente_hasta": {
            "type": "string",
            "format": "date-time",
            "title": "Vigente Hasta"
          }
        },
        "type": "object",
        "required": [
          "id",
          "titulo",
          "resumen",
          "enlace",
          "fuente",
          "region_dane",
          "cultivo_id",
          "publicada",
          "vigente_hasta"
        ],
        "title": "NoticiaSalida"
      },
...
      "Page_NoticiaSalida_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/NoticiaSalida"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "size",
          "has_more"
        ],
        "title": "Page[NoticiaSalida]"
      },
...
      "Page_PropagacionSalida_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PropagacionSalida"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "size",
          "has_more"
        ],
        "title": "Page[PropagacionSalida]"
      },
...
      "PropagacionActualizar": {
        "properties": {
          "germinadas": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Germinadas"
          },
          "listas": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Listas"
          },
          "perdidas": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Perdidas"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "PropagacionActualizar"
      },
      "PropagacionCrear": {
        "properties": {
          "finca_id": {
            "type": "string",
            "format": "uuid",
            "title": "Finca Id"
          },
          "cultivo_id": {
            "type": "string",
            "format": "uuid",
            "title": "Cultivo Id"
          },
          "metodo": {
            "type": "string",
            "enum": [
              "semilla",
              "esqueje",
              "estaca",
              "injerto",
              "hijuelo",
              "acodo",
              "in_vitro"
            ],
            "title": "Metodo"
          },
          "fecha_inicio": {
            "type": "string",
            "format": "date",
            "title": "Fecha Inicio"
          },
          "puestas": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Puestas"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "finca_id",
          "cultivo_id",
          "metodo",
          "fecha_inicio",
          "puestas"
        ],
        "title": "PropagacionCrear"
      },
      "PropagacionSalida": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "finca_id": {
            "type": "string",
            "format": "uuid",
            "title": "Finca Id"
          },
          "cultivo_id": {
            "type": "string",
            "format": "uuid",
            "title": "Cultivo Id"
          },
          "siembra_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Siembra Id"
          },
          "metodo": {
            "type": "string",
            "title": "Metodo"
          },
          "fecha_inicio": {
            "type": "string",
            "format": "date",
            "title": "Fecha Inicio"
          },
          "puestas": {
            "type": "integer",
            "title": "Puestas"
          },
          "germinadas": {
            "type": "integer",
            "title": "Germinadas"
          },
          "listas": {
            "type": "integer",
            "title": "Listas"
          },
          "perdidas": {
            "type": "integer",
            "title": "Perdidas"
          },
          "trasplantadas": {
            "type": "integer",
            "title": "Trasplantadas"
          }
        },
        "type": "object",
        "required": [
          "id",
          "finca_id",
          "cultivo_id",
          "siembra_id",
          "metodo",
          "fecha_inicio",
          "puestas",
          "germinadas",
          "listas",
          "perdidas",
          "trasplantadas"
        ],
        "title": "PropagacionSalida"
      },
...
      "TrasplanteEntrada": {
        "properties": {
          "siembra_id": {
            "type": "string",
            "format": "uuid",
            "title": "Siembra Id"
          },
          "cantidad": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Cantidad"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "siembra_id",
          "cantidad"
        ],
        "title": "TrasplanteEntrada"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  },
...
}